Linear search using loops

+++ main.cpp
@@ -1,10 +1,11 @@
 #include<iostream>
+#include<climits>
 using namespace std;
 int main()
 {
     int nums[]={10,20,45};
     int size=3;
-    int smallest=INT8_MAX;
+    int smallest=INT_MAX;
 for(int i=0; i<size; i++)
 { 
     if(nums[i]<smallest)
